+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- .gitattributes file
+
 ## [2.0.1] - (05/02/2020)
 
 ### Added